contract_intelligence/pyproject.toml
[project]
name = "contract-intelligence"
version = "0.1.0"
description = "Hybrid Contract Intelligence Platform using GraphRAG and PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "asyncpg",
    "pgvector",
    "openai==2.8.1",
    "azure-identity",
    "graphrag",
    "agent-framework==1.0.0b251120",
    "PyPDF2",
    "pdf2image",
    "pytesseract",
    "docling",
    "Pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
dev-dependencies = []
